Adding files from old gasdospela project
If git preserves symlinks one should be able to compile everything
with a simple 'make'. Otherwise just do:

  for f in ../kens-kode/*.s??; do ln -s $f; done

From 'jespers-kode'.

The Analysis.sml is the hacked up version of our initial
analysis-framework. I have beat it until a function core2gcl_new
(great name btw) appeared.

The must somehow be an error in the translation of preconditions in
Core where two variables are compared because for the VerySimple.fsm
example, the translation to Gcl and subsequent BDD for transition
relation seem to be correct the script is only comparing variables (in
Core) to constants.