Modify translation of commands to use AND instead of IMP
I somehow think translating p?us into [p] => [us] is wrong and below
is an example demonstrating why I think so.

(sorry that the comnmit message is just rambling on...)

Gcl:

a ? b:=T
NOT(a) ? c:=T

The translation is then easy (simply replace ? with IMP). The
translation using IMP would correspond to the boolean formula:

a => b'  +  not(a) => c'

I think that the truthtable of the formula shows that the formula is
valid and thus the bdd representing it would be the 1-bdd.

I guess it is formally correct that the transitions of the mentioned
Gcl program can be modelled by the 1-bdd. Since there are 3 variables
(a, b, and c) we simply have to make sure that the bdd (boolean
function) representing the transitions says that :
- whenever 'a' is true, 'b' should be set to true irrespective of 'c's
  and 'b's current value, and
- whenever 'a' is not true, 'c' should be set to true irrespective of
  'c's and 'b's current value.

If we use the 'AND' translation instead of 'IMP' of Gcl commands, we
get instead the boolean formula:

a & b' + not(a) & c'

Certainly, this is satisfieable but not valid and so not represented
by the 1-bdd.

Anyway, we get the a => ... + not(a) => ... in the Gcl program exactly
because of how I translate preconditions. Maybe they could simply be
done better?